+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
-import Header from './Header.js';
-import Home from './Home.js';
+import Header from './Header/Header.js';
+import Home from './Products/HomePage/Home.js';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import '@aws-amplify/ui-react/styles.css';
